+++ src/core/server/saved_objects/import/resolve_import_errors.ts
@@ -61,6 +61,7 @@ export async function resolveSavedObjectsImportErrors({
   createNewCopies,
   dataSourceId,
   dataSourceTitle,
+  workspaces,
 }: SavedObjectsResolveImportErrorsOptions): Promise<SavedObjectsImportResponse> {
   // throw a BadRequest error if we see invalid retries
   validateRetries(retries);
@@ -163,6 +164,7 @@ export async function resolveSavedObjectsImportErrors({
       overwrite,
       dataSourceId,
       dataSourceTitle,
+      workspaces,
     };
     const { createdObjects, errors: bulkCreateErrors } = await createSavedObjects(
       createSavedObjectsParams